# Improve reliability and performance of GL disposals