Resolves #4108.

To quote my comment there:

> Commit messages and inline comments should hopefully explain the main idea. From testing:
>
> * It passes the 100k drawable stress test given above.
> * I've benchmarked it on the test using dotTrace. Also tried adding a Benchmark.NET, but it's not trivial as it requires an actual game host and measuring that would probably be useless. It doesn't seem to be worse with the reworked concept, and it certainly seems to be better with the optimisation (times given are total times of the appropriate operations, measurement was once each time, over 2 minutes of running the stress test):
>
> | | current `master` | with `GLDisposalQueue` | with optimised `GLDisposalQueue` |
> | :- | -: | -: | -: |
> | scheduling disposals | 342ms | 311ms | 301ms |
> | checking disposals | 2267ms | 2179ms | 622ms |

I've also benchmarked against the game using dotTrace. Testing procedure was: release build, one-time run, from intro, through one map, to game exit - around 4 mins of operation overall each time. The results there are less impressive, but maybe that was to be expected, as the stress test case was a pretty degenerate one. Table for reference:

| | current `master` | with `GLDisposalQueue` | with optimised `GLDisposalQueue` |
| :- | -: | -: | -: |
| scheduling disposals | 19ms | 22ms | 9.3ms |
| checking disposals | 473ms | 744ms | 394ms |

That middle result might be an artifact of thermal throttling on my box, but I don't think it's all that relevant of a result as the optimisation in 5971ae2 seems like a safe optimisation to make (previous code was operating off of a similar assumption anyway - see the `Debug.Assert(i == 0)` before removal), and it's not like it's a super jarring difference - might just be normal-ish fluctuation.